[stdlib] Refactor and Unify Token (#885)
1. Allow anyone to burn STC.
2. Refactor Token value from u64 to u128.
3. Clean and refactor Token/Coin's Rust types.
4. Set arbitrary_precision serde_json features to support json rpc u128.
5. Bump miner jsonrpc to 14.0 for enabling json arbitrary_precision feature.
6. enable HeadBlockPacemaker in dev network.

+++ vm/stdlib/init_scripts/genesis_init.move
@@ -7,7 +7,7 @@ fun genesis_init(publishing_option: vector<u8>, instruction_schedule: vector<u8>
                  reward_half_epoch: u64, init_block_time_target: u64,
                  block_difficulty_window: u64, reward_per_uncle_percent: u64,
                  min_time_target:u64, max_uncles_per_block:u64,
-                 total_supply: u64, pre_mine_percent:u64, parent_hash: vector<u8>,
+                 total_supply: u128, pre_mine_percent:u64, parent_hash: vector<u8>,
                  association_auth_key: vector<u8>, genesis_auth_key: vector<u8>,
                  ) {
     Genesis::initialize(publishing_option, instruction_schedule,
