+++ vm/stdlib/transaction_scripts/create_account.move
@@ -1,8 +1,7 @@
 script {
 use 0x1::Account;
 
-// TODO: remove initial_amount?
-fun main<TokenType>(account: &signer, fresh_address: address, auth_key_prefix: vector<u8>, initial_amount: u64) {
+fun main<TokenType>(account: &signer, fresh_address: address, auth_key_prefix: vector<u8>, initial_amount: u128) {
   Account::create_account<TokenType>(fresh_address, auth_key_prefix);
   if (initial_amount > 0) Account::deposit(account,
         fresh_address,
